+++ weaver/datatype.py
@@ -210,6 +210,7 @@ class Base(DictBase):
     """
     Base interface for all data-types.
     """
+
     def __str__(self):
         # type: () -> str
         return f"{type(self).__name__} <{self.id}>"
@@ -255,6 +256,7 @@ class LocalizedDateTimeProperty(property):
     """
     Property that ensures date-time localization is applied on the stored/retrieved value as required.
     """
+
     def __init__(self,
                  fget=None,             # type: Callable[[Any], Optional[datetime]]
                  fset=None,             # type: Callable[[Any, Union[datetime, str]], None]